+++ lib/default.nix
@@ -1,5 +1,6 @@
 { pkgs, lib }:
 {
   mkHermesAgent = import ./mkHermesAgent.nix { inherit pkgs lib; };
+  mkZeroClawAgent = import ./mkZeroClawAgent.nix { inherit pkgs lib; };
   constants = import ./constants.nix;
 }